+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 This changelog is for older versions. See [Github Releases](https://github.com/inovait/neatle/releases) for newer versions.
 
+## Version 0.10.0
+ - New: Added setTransport() to Connection
+ - New: Added setTransport() to ConnectionMonitor
+
 ## Version 0.9.3
  - Fixed: #21 - BluetoothGattCharacteristic needed to extract value on subscription